Use default gcc if version >= 4.9
On ubuntu 16.04, the default /usr/bin/g{cc,++} compiler is used to
build SciDB. The c++ ABI chaged with the libstdc++ library has changed
for g++ 5.  Since SciDB uses the NEW ABI, the compilation of shim
with the older g++-4.9 is unable to use the symbols from the New ABI.
(That is, the libstdc++ is backwards compatible, but the libstdc++
used by gcc 4.9 is not forward compatible)

g++ (Ubuntu 5.4.0-6ubuntu1~16.04.12) 5.4.0 20160609
Copyright (C) 2015 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.